MT_PointCachebatch Updated
Was previously not working, this was found to be due to a conflict with the now no longer used MAKE_PointCache modifier.

By removing the MAKE_PointCache modifier from the plugins the functions now work correctly.

It appears that there is a named function or variable shared between the MAKE_PointCache and the default max one in this build of 3dsmax. Not that surprising considering MAKE might be a common prefix on an action.

This issue was found when running the cacheOps.RecordCache function, causing a memory access violation.

I also cleaned up some baseball happening to make the reporting more visible.

Co-Authored-By: TheOnlyAaron <theonlyaaron@gmail.com>